+++ healthcheck/__init__.py
@@ -1,9 +1,8 @@
 __doc__ = 'Health Checker for Yola Services'
-__version__ = '0.0.3'
+__version__ = '0.0.4'
 __url__ = 'https://github.com/yola/healthcheck'
 
 
 from healthcheck import (HealthChecker, HealthCheck, ListHealthCheck,
                          DjangoDBsHealthCheck, FilesExistHealthCheck,
                          FilesDontExistHealthCheck)
-